Fixed bug that caused visual crash upon attempting to change search terms. Turned on the option "noImplicitAny" in tsConfig, and fixed all resulting errors
The function updater at app\javascript\components\SearchPanel.tsx:98 expects type (tag:string[]) => string[].

However, I passed it the object searchList[tabID], where searchList is of the type {tabDataMap: Map<number,SearchTabDataProp>, tabOrder: number[]},
and tabID is of type numeric.

What I expect is for it to throw up an error, but what i got is the type any.

Debugging gives me this link: https://bartvds.github.io/tsspec-html/0.9.1/dist/expressions.html
Relevant Sections is 4.10 Property Access, on "bracket notation property access"

Quote:

A bracket notation property access of the form

        ObjExpr [ IndexExpr ]

where ObjExpr and IndexExpr are expressions, is used to access the property with the name computed by the index expression on the given object.
A bracket notation property access is processed as follows at compile-time:

If IndexExpr is a string literal or a numeric literal and ObjExpr is of a type that has a property with the name given by that literal                (converted to its string representation in the case of a numeric literal), the property access is of the type of that property.

End Quote
The final diagnosis is that this counted as a dynamic property access, which was assigned a type any.
As a result, I turned on the option "noImplicitAny" in tsConfig, to explictly flag such warnings